+++ Modules/Microsoft365DSC/DscResources/MSFT_EXOManagementRoleEntry/MSFT_EXOManagementRoleEntry.psm1
@@ -59,6 +59,12 @@ function Get-TargetResource
 
     Write-Verbose -Message "Getting Management Role Entry configuration for {$Identity}"
 
+    if ($PSBoundParameters.ContainsKey('Type'))
+    {
+        $PSBoundParameters.Remove('Type') | Out-Null
+        Write-Warning "Property 'Type' is deprecated and will be removed“
+    }
+
     try
     {
         if (-not $Script:exportedInstance -or $Script:exportedInstance.Identity -ne $Identity)
@@ -180,6 +186,12 @@ function Set-TargetResource
     )
 
     Write-Verbose -Message "Setting Management Role Entry configuration for {$Identity}"
+
+    if ($PSBoundParameters.ContainsKey('Type'))
+    {
+        $PSBoundParameters.Remove('Type') | Out-Null
+        Write-Warning "Property 'Type' is deprecated and will be removed“
+    }
 
     #Ensure the proper dependencies are installed in the current environment.
     Confirm-M365DSCDependencies
@@ -316,8 +328,10 @@ function Test-TargetResource
     Add-M365DSCTelemetryEvent -Data $data
     #endregion
 
+    $compareParameters = Get-CompareParameters
     $result = Test-M365DSCTargetResource -DesiredValues $PSBoundParameters `
-        -ResourceName $($MyInvocation.MyCommand.Source).Replace('MSFT_', '')
+        -ResourceName $($MyInvocation.MyCommand.Source).Replace('MSFT_', '') `
+        @compareParameters
     return $result
 }
 
@@ -435,4 +449,15 @@ function Export-TargetResource
     }
 }
 
-Export-ModuleMember -Function *-TargetResource
+function Get-CompareParameters
+{
+    [CmdletBinding()]
+    [OutputType([System.Collections.Hashtable])]
+    param()
+
+    return @{
+        ExcludedProperties = @('Type')
+    }
+}
+
+Export-ModuleMember -Function @('*-TargetResource', 'Get-CompareParameters')